Move dataset tests to separate file
